#9 removed using of DI in tests combined configuration tests in one class
