+++ src/test/java/tests/configurations/ProfileConfigurationTests.java
@@ -1,7 +1,8 @@
 package tests.configurations;
 
 import aquality.selenium.core.application.AqualityModule;
-import aquality.selenium.core.configurations.ILoggerConfiguration;
+import aquality.selenium.core.configurations.LoggerConfiguration;
+import aquality.selenium.core.utilities.ISettingsFile;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import tests.application.CustomAqualityServices;
@@ -20,7 +21,7 @@ public class ProfileConfigurationTests extends BaseProfileTest {
 
     @Test
     public void testShouldBePossibleToGetLanguageFromNewSettingsFile() {
-        String language = CustomAqualityServices.getServiceProvider().getInstance(ILoggerConfiguration.class).getLanguage();
+        String language = new LoggerConfiguration(CustomAqualityServices.getServiceProvider().getInstance(ISettingsFile.class)).getLanguage();
         assertEquals(language, "ru", String.format("Current language should be got from %s profile", PROFILE));
     }
 }
