+++ src/test/resources/TestSuite.xml
@@ -6,10 +6,7 @@
             <class name="tests.application.AqualityServicesTests"/>
             <class name="tests.application.browser.BrowserTests"/>
             <class name="tests.utilities.CustomSettingsFileTests"/>
-            <class name="tests.configurations.LoggerConfigurationTests"/>
-            <class name="tests.configurations.TimeoutConfigurationTests"/>
-            <class name="tests.configurations.RetryConfigurationTests"/>
-            <class name="tests.configurations.ElementCacheConfigurationTests"/>
+            <class name="tests.configurations.ConfigurationTests"/>
             <class name="tests.utilities.ElementActionRetrierTests"/>
         </classes>
     </test>
